[noissue] Extract shared MockAgentService and mock helpers to reduce test duplication
Consolidate ~1,200 lines of duplicated MockAgentService, MockHeaderReader,
MockTraceContextReader/Writer, and MockCallStackReader definitions from 8 test
files into two shared headers (mock_agent_service.h, mock_helpers.h).
