+++ test/test_noop.cpp
@@ -27,204 +27,29 @@
 #include "../src/url_stat.h"
 #include "../src/stat.h"
 #include "../include/pinpoint/tracer.h"
+#include "mock_agent_service.h"
+#include "mock_helpers.h"
 
 namespace pinpoint {
-
-class MockAgentService : public AgentService {
-public:
-    MockAgentService() : is_exiting_(false), start_time_(std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count()), cached_start_time_str_(std::to_string(start_time_)), trace_id_counter_(0) {
-        config_->span.event_chunk_size = 10; // Set a reasonable default for testing
-        config_->span.max_event_depth = 32;
-        config_->span.queue_size = 1024;
-        config_->http.url_stat.enable = true; // Enable URL stat for relevant tests
-        config_->http.url_stat.limit = 1024;
-        config_->http.url_stat.trim_path_depth = 3;
-    }
-
-    bool isExiting() const override { return is_exiting_; }
-    void setExiting(bool exiting) { is_exiting_ = exiting; }
-
-    std::string getAppName() const override { return "mock-app"; }
-    int32_t getAppType() const override { return 1300; }
-    std::string getAgentId() const override { return "mock-agent"; }
-    std::string getAgentName() const override { return "mock-agent-name"; }
-    std::shared_ptr<const Config> getConfig() const override { return config_; }
-    int64_t getStartTime() const override { return start_time_; }
-    void reloadConfig(std::shared_ptr<const Config> cfg) override {
-        if (cfg) {
-            *config_ = *cfg;
-        }
-    }
-
-    TraceId generateTraceId() override {
-        return TraceId{"mock-agent", start_time_, trace_id_counter_++};
-    }
-    void recordSpan(std::unique_ptr<SpanChunk> span) const override {
-        recorded_spans_.push_back(std::move(span));
-    }
-    void recordUrlStat(std::unique_ptr<UrlStatEntry> stat) const override {
-        recorded_url_stats_++;
-        last_url_stat_url_ = stat->url_pattern_;
-        last_url_stat_method_ = stat->method_;
-        last_url_stat_status_code_ = stat->status_code_;
-    }
-    void recordException(SpanData* span_data) const override {
-        recorded_exceptions_++;
-    }
-    void recordStats(StatsType stats) const override {
-        recorded_stats_calls_++;
-        last_stats_type_ = stats;
-    }
-
-    int32_t cacheApi(std::string_view api_str, int32_t api_type) const override {
-        if (cached_apis_.find(std::string(api_str)) == cached_apis_.end()) {
-            cached_apis_[std::string(api_str)] = api_id_counter_++;
-        }
-        return cached_apis_[std::string(api_str)];
-    }
-    void removeCacheApi(const ApiMeta& api_meta) const override {}
-    int32_t cacheError(std::string_view error_name) const override {
-        if (cached_errors_.find(std::string(error_name)) == cached_errors_.end()) {
-            cached_errors_[std::string(error_name)] = error_id_counter_++;
-        }
-        return cached_errors_[std::string(error_name)];
-    }
-    void removeCacheError(const StringMeta& error_meta) const override {}
-
-    int32_t cacheSql(std::string_view sql_query) const override {
-        if (cached_sqls_.find(std::string(sql_query)) == cached_sqls_.end()) {
-            cached_sqls_[std::string(sql_query)] = sql_id_counter_++;
-        }
-        return cached_sqls_[std::string(sql_query)];
-    }
-    void removeCacheSql(const StringMeta& sql_meta) const override {}
-
-    std::vector<unsigned char> cacheSqlUid(std::string_view sql) const override {
-        // Mock implementation - return test uid
-        return {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
-    }
-
-    void removeCacheSqlUid(const SqlUidMeta& sql_uid_meta) const override {
-        // Mock implementation
-    }
-
-    bool isStatusFail(int status) const override { return status >= 400; }
-    void recordServerHeader(HeaderType which, HeaderReader& reader, const AnnotationPtr& annotation) const override {
-        recorded_server_headers_++;
-    }
-    void recordClientHeader(HeaderType which, HeaderReader& reader, const AnnotationPtr& annotation) const override {
-        recorded_client_headers_++;
-    }
-
-    AgentStats& getAgentStats() override {
-        if (!agent_stats_) {
-            agent_stats_ = std::make_unique<AgentStats>(this);
-        }
-        return *agent_stats_;
-    }
-
-    UrlStats& getUrlStats() override {
-        if (!url_stats_) {
-            url_stats_ = std::make_unique<UrlStats>(this);
-        }
-        return *url_stats_;
-    }
-
-    // Test-specific accessors
-    mutable int recorded_stats_calls_ = 0;
-    mutable StatsType last_stats_type_ = AGENT_STATS;
-    mutable int recorded_url_stats_ = 0;
-    mutable int recorded_exceptions_ = 0;
-    mutable std::string last_url_stat_url_;
-    mutable std::string last_url_stat_method_;
-    mutable int last_url_stat_status_code_ = 0;
-    mutable int recorded_server_headers_ = 0;
-    mutable int recorded_client_headers_ = 0;
-    mutable std::vector<std::unique_ptr<SpanChunk>> recorded_spans_;
-    mutable std::map<std::string, int32_t> cached_apis_;
-    mutable std::map<std::string, int32_t> cached_errors_;
-    mutable std::map<std::string, int32_t> cached_sqls_;
-    mutable int32_t api_id_counter_ = 100;
-    mutable int32_t error_id_counter_ = 200;
-    mutable int32_t sql_id_counter_ = 300;
-
-private:
-    bool is_exiting_;
-    int64_t start_time_;
-    std::string cached_start_time_str_;
-    int64_t trace_id_counter_;
-    std::shared_ptr<Config> config_ = std::make_shared<Config>();
-    mutable std::unique_ptr<AgentStats> agent_stats_;
-    mutable std::unique_ptr<UrlStats> url_stats_;
-};
-
-class MockTraceContextWriter : public TraceContextWriter {
-public:
-    MockTraceContextWriter() = default;
-    void Set(std::string_view key, std::string_view value) override {
-        context_[std::string(key)] = std::string(value);
-    }
-    std::optional<std::string> Get(std::string_view key) const {
-        auto it = context_.find(std::string(key));
-        if (it != context_.end()) {
-            return it->second;
-        }
-        return std::nullopt;
-    }
-private:
-    std::map<std::string, std::string> context_;
-};
-
-class MockTraceContextReader : public TraceContextReader {
-public:
-    MockTraceContextReader() = default;
-    std::optional<std::string> Get(std::string_view key) const override {
-        auto it = context_.find(std::string(key));
-        if (it != context_.end()) {
-            return it->second;
-        }
-        return std::nullopt;
-    }
-    void SetContext(std::string_view key, std::string_view value) {
-        context_[std::string(key)] = std::string(value);
-    }
-private:
-    std::map<std::string, std::string> context_;
-};
-
-class MockHeaderReader : public HeaderReader {
-public:
-    MockHeaderReader() = default;
-
-    std::optional<std::string> Get(std::string_view key) const override {
-        auto it = headers_.find(std::string(key));
-        if (it != headers_.end()) {
-            return it->second;
-        }
-        return std::nullopt;
-    }
-
-    void ForEach(std::function<bool(std::string_view key, std::string_view val)> callback) const override {
-        for (const auto& pair : headers_) {
-            if (!callback(pair.first, pair.second)) {
-                break;
-            }
-        }
-    }
-
-    void SetHeader(std::string_view key, std::string_view value) {
-        headers_[std::string(key)] = std::string(value);
-    }
-
-private:
-    std::map<std::string, std::string> headers_;
-};
 
 class NoopTest : public ::testing::Test {
 protected:
     void SetUp() override {
         mock_agent_service_ = std::make_unique<MockAgentService>();
         mock_agent_service_->setExiting(false);
+        mock_agent_service_->setAppName("mock-app");
+        mock_agent_service_->setAgentId("mock-agent");
+        mock_agent_service_->setAgentName("mock-agent-name");
+        mock_agent_service_->setStartTime(
+            std::chrono::duration_cast<std::chrono::milliseconds>(
+                std::chrono::system_clock::now().time_since_epoch()).count());
+        auto& cfg = mock_agent_service_->mutableConfig();
+        cfg->span.event_chunk_size = 10;
+        cfg->span.max_event_depth = 32;
+        cfg->span.queue_size = 1024;
+        cfg->http.url_stat.enable = true;
+        cfg->http.url_stat.limit = 1024;
+        cfg->http.url_stat.trim_path_depth = 3;
     }
 
     void TearDown() override {
@@ -459,17 +284,11 @@ TEST_F(NoopTest, NoopSpanEventAllMethodsTest) {
     SUCCEED() << "All NoopSpanEvent methods should execute without throwing exceptions";
 }
 
-class MockCallStackReader : public CallStackReader {
-public:
-    void ForEach(std::function<void(std::string_view module, std::string_view function, std::string_view file, int line)> callback) const override {
-        callback("test-module", "test_function", "test.cpp", 42);
-    }
-};
-
 TEST_F(NoopTest, NoopSpanEventSetErrorWithCallStackTest) {
     NoopSpanEvent span_event;
 
     MockCallStackReader stack_reader;
+    stack_reader.AddFrame("test-module", "test_function", "test.cpp", 42);
     span_event.SetError("TestError", "Test error message", stack_reader);
 
     SUCCEED() << "SetError with CallStackReader should execute without throwing";
